requirements.txt
numpy==1.20.3
torch==1.9.0
torchvision==0.10.0
streamlit==0.88.0
matplotlib==3.4.2
streamlit_drawable_canvas==0.8.0
scipy==1.6.2
Pillow==8.3.2
